fix vaadin latest dep test
