+++ instrumentation/vaadin-14.2/javaagent/src/vaadin16Test/java/io/opentelemetry/javaagent/instrumentation/vaadin/Vaadin16Test.java
@@ -5,13 +5,67 @@
 
 package io.opentelemetry.javaagent.instrumentation.vaadin;
 
-import java.io.File;
+import static org.assertj.core.api.Assertions.assertThat;
+import static org.awaitility.Awaitility.await;
 
-public class Vaadin16Test extends AbstractVaadin16Test {
+import io.opentelemetry.api.trace.SpanKind;
+import io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions;
+import io.opentelemetry.sdk.testing.assertj.TracesAssert;
+import io.opentelemetry.sdk.trace.data.SpanData;
+import java.io.File;
+import java.util.List;
+
+public class Vaadin16Test extends AbstractVaadinTest {
 
   @Override
   protected void prepareVaadinBaseDir(File baseDir) {
     copyResource("/pnpm/package.json", baseDir);
     copyResource("/pnpm/pnpm-lock.yaml", baseDir);
+  }
+
+  @Override
+  void assertFirstRequest() {
+    await()
+        .untilAsserted(
+            () -> {
+              List<List<SpanData>> traces = testing.waitForTraces(1);
+              TracesAssert.assertThat(traces)
+                  .satisfies(
+                      trace -> {
+                        assertThat(trace.get(0))
+                            .satisfies(
+                                spans ->
+                                    OpenTelemetryAssertions.assertThat(spans.get(0))
+                                        .hasName("GET IndexHtmlRequestHandler.handleRequest")
+                                        .hasNoParent()
+                                        .hasKind(SpanKind.SERVER));
+                        assertThat(trace)
+                            .anySatisfy(
+                                spans -> {
+                                  OpenTelemetryAssertions.assertThat(spans.get(0))
+                                      .hasName("POST " + getContextPath() + "/main")
+                                      .hasNoParent()
+                                      .hasKind(SpanKind.SERVER);
+                                  OpenTelemetryAssertions.assertThat(spans.get(1))
+                                      .hasName("SpringVaadinServletService.handleRequest")
+                                      .hasParent(spans.get(0))
+                                      .hasKind(SpanKind.INTERNAL);
+                                  // we don't assert all the handler spans as these vary between
+                                  // vaadin versions
+                                  OpenTelemetryAssertions.assertThat(spans.get(spans.size() - 3))
+                                      .hasName("UidlRequestHandler.handleRequest")
+                                      .hasParent(spans.get(1))
+                                      .hasKind(SpanKind.INTERNAL);
+                                  OpenTelemetryAssertions.assertThat(spans.get(spans.size() - 2))
+                                      .hasName("PublishedServerEventHandlerRpcHandler.handle")
+                                      .hasParent(spans.get(spans.size() - 3))
+                                      .hasKind(SpanKind.INTERNAL);
+                                  OpenTelemetryAssertions.assertThat(spans.get(spans.size() - 1))
+                                      .hasName("JavaScriptBootstrapUI.connectClient")
+                                      .hasParent(spans.get(spans.size() - 2))
+                                      .hasKind(SpanKind.INTERNAL);
+                                });
+                      });
+            });
   }
 }
