--- instrumentation/vaadin-14.2/testing/src/main/java/io/opentelemetry/javaagent/instrumentation/vaadin/AbstractVaadin16Test.java
@@ -1,63 +1,0 @@
-/*
- * Copyright The OpenTelemetry Authors
- * SPDX-License-Identifier: Apache-2.0
- */
-
-package io.opentelemetry.javaagent.instrumentation.vaadin;
-
-import static org.assertj.core.api.Assertions.assertThat;
-import static org.awaitility.Awaitility.await;
-
-import io.opentelemetry.api.trace.SpanKind;
-import io.opentelemetry.sdk.testing.assertj.OpenTelemetryAssertions;
-import io.opentelemetry.sdk.testing.assertj.TracesAssert;
-import io.opentelemetry.sdk.trace.data.SpanData;
-import java.util.List;
-
-public class AbstractVaadin16Test extends AbstractVaadinTest {
-  @Override
-  void assertFirstRequest() {
-    await()
-        .untilAsserted(
-            () -> {
-              List<List<SpanData>> traces = testing.waitForTraces(1);
-              TracesAssert.assertThat(traces)
-                  .satisfies(
-                      trace -> {
-                        assertThat(trace.get(0))
-                            .satisfies(
-                                spans ->
-                                    OpenTelemetryAssertions.assertThat(spans.get(0))
-                                        .hasName("GET IndexHtmlRequestHandler.handleRequest")
-                                        .hasNoParent()
-                                        .hasKind(SpanKind.SERVER));
-                        assertThat(trace)
-                            .anySatisfy(
-                                spans -> {
-                                  OpenTelemetryAssertions.assertThat(spans.get(0))
-                                      .hasName("POST " + getContextPath() + "/main")
-                                      .hasNoParent()
-                                      .hasKind(SpanKind.SERVER);
-                                  OpenTelemetryAssertions.assertThat(spans.get(1))
-                                      .hasName("SpringVaadinServletService.handleRequest")
-                                      .hasParent(spans.get(0))
-                                      .hasKind(SpanKind.INTERNAL);
-                                  // we don't assert all the handler spans as these vary between
-                                  // vaadin versions
-                                  OpenTelemetryAssertions.assertThat(spans.get(spans.size() - 3))
-                                      .hasName("UidlRequestHandler.handleRequest")
-                                      .hasParent(spans.get(1))
-                                      .hasKind(SpanKind.INTERNAL);
-                                  OpenTelemetryAssertions.assertThat(spans.get(spans.size() - 2))
-                                      .hasName("PublishedServerEventHandlerRpcHandler.handle")
-                                      .hasParent(spans.get(spans.size() - 3))
-                                      .hasKind(SpanKind.INTERNAL);
-                                  OpenTelemetryAssertions.assertThat(spans.get(spans.size() - 1))
-                                      .hasName("JavaScriptBootstrapUI.connectClient")
-                                      .hasParent(spans.get(spans.size() - 2))
-                                      .hasKind(SpanKind.INTERNAL);
-                                });
-                      });
-            });
-  }
-}
